libras-recognizer/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
ultralytics>=8.0.0
pillow>=8.3.1
numpy>=1.21.2
opencv-python-headless>=4.5.0
python-multipart>=0.0.6